origami_ocr/requirements.txt
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.10
click==8.0.3
cloudpickle==2.0.0
configparser==5.2.0
docker-pycreds==0.4.0
editdistance==0.6.0
gin-config==0.5.0
gitdb==4.0.9
GitPython==3.1.26
idna==3.3
imageio==2.14.0
joblib==1.1.0
networkx==2.6.3
nltk==3.6.7
numpy==1.22.1
opencv-python==4.5.5.62
packaging==21.3
pathtools==0.1.2
Pillow==9.0.0
promise==2.3
protobuf==3.19.3
psutil==5.9.0
pycparser==2.21
pyparsing==3.0.7
python-dateutil==2.8.2
PyWavelets==1.2.0
PyYAML==6.0
regex==2022.1.18
requests==2.27.1
scikit-image==0.19.1
scipy==1.7.3
sentry-sdk==1.5.3
shortuuid==1.0.8
six==1.16.0
smmap==5.0.0
subprocess32==3.5.4
termcolor==1.1.0
tifffile==2021.11.2
torch==1.10.1
torchvision==0.11.2
tqdm==4.62.3
typing_extensions==4.0.1
urllib3==1.26.8
wandb==0.12.9
yaspin==2.1.0
